Move TestTrace to tracing crate
