+++ crates/tracing/src/lib.rs
@@ -16,3 +16,18 @@ impl CanTrace for Lib3hTrace {
         unimplemented!()
     }
 }
+
+#[derive(Debug)]
+pub struct TestTrace(pub String);
+
+impl CanTrace for TestTrace {
+    fn get_span(&self) -> Span {
+        unimplemented!()
+    }
+}
+
+impl TestTrace {
+    pub fn new() -> Self {
+        Self("Unnamed TestTrace".into())
+    }
+}
